Rename .es files to .js files since they are already in ES folder
